Fix Error: src/cscaledbitmap.cxx:456:26: error: result of comparison of constant 256 with expression of type 'uint8_t' (aka 'unsigned char') is always true

+++ graphics/nxwidgets/src/cscaledbitmap.cxx
@@ -433,7 +433,7 @@ bool CScaledBitmap::scaleColor(FAR const struct rgbcolor_s &incolor1,
                                FAR const struct rgbcolor_s &incolor2,
                                b16_t fraction, FAR struct rgbcolor_s &outcolor)
 {
-  uint8_t component;
+  uint32_t component;
   b16_t red;
   b16_t green;
   b16_t blue;
